+++ src/main/scala/gwen/core/Predefs.scala
@@ -258,14 +258,14 @@ object Formatting {
   def padTailLines(str: String, padding: String) = str.replaceAll("""\r?\n""", s"""\n$padding""")
   def sha256Hash(source: String): String = DigestUtils.sha256Hex(source)
 
-  def formatTable(table: List[(Int, List[String])]): String = {
+  def formatTable(table: List[(Long, List[String])]): String = {
     (table.indices.toList map { rowIndex => formatTableRow(table, rowIndex) }).mkString("\r\n")
   }
-  def formatTableRow(table: List[(Int, List[String])], rowIndex: Int): String = {
+  def formatTableRow(table: List[(Long, List[String])], rowIndex: Int): String = {
     val maxWidths = (table map { case (_, rows) => rows.map(_.length) }).transpose.map(_.max)
     s"| ${(table(rowIndex)._2.zipWithIndex map { case (data, dataIndex) => s"${rightPad(data, maxWidths(dataIndex))}" }).mkString(" | ") } |"
   }
-  def formatDocString(docString: (Int, String, Option[String]), includeType: Boolean = true) = docString match {
+  def formatDocString(docString: (Long, String, Option[String]), includeType: Boolean = true) = docString match {
     case (_, content, contentType) =>
       s"""|${"\"\"\""}${if(includeType) contentType.getOrElse("") else ""}
           |$content
